@@ -1,34 +1,48 @@
 {
 	"name": "svelte",
-	"version": "0.0.1",
 	"private": true,
+	"version": "0.0.1",
+	"type": "module",
 	"scripts": {
 		"dev": "vite dev",
 		"build": "vite build",
 		"preview": "vite preview",
-		"lint": "prettier --plugin-search-dir . --check . && eslint .",
-		"format": "prettier --plugin-search-dir . --write ."
-	},
-	"type": "module",
-	"dependencies": {
-		"@sanity/client": "^6.19.1",
-		"@sanity/image-url": "^1.0.2",
-		"@sveltejs/vite-plugin-svelte": "^3.1.1",
-		"uuid": "^9.0.1"
+		"prepare": "svelte-kit sync || echo ''",
+		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
+		"format": "prettier --write .",
+		"lint": "prettier --check . && eslint ."
 	},
 	"devDependencies": {
-		"@portabletext/svelte": "^2.1.11",
-		"@sveltejs/adapter-auto": "^3.2.1",
-		"@sveltejs/kit": "^2.5.10",
-		"autoprefixer": "^10.4.19",
-		"eslint": "^9.4.0",
-		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.39.0",
-		"postcss": "^8.4.38",
-		"prettier": "^3.3.0",
-		"prettier-plugin-svelte": "^3.2.3",
-		"svelte": "^4.2.17",
-		"tailwindcss": "^3.4.3",
-		"vite": "^5.2.12"
+		"@eslint/compat": "^1.2.5",
+		"@eslint/js": "^9.18.0",
+		"@portabletext/svelte": "^3.0.0",
+		"@sveltejs/adapter-auto": "^4.0.0",
+		"@sveltejs/kit": "^2.16.0",
+		"@sveltejs/vite-plugin-svelte": "^5.0.0",
+		"@tailwindcss/vite": "^4.0.0",
+		"eslint": "^9.18.0",
+		"eslint-config-prettier": "^10.0.1",
+		"eslint-plugin-svelte": "^3.0.0",
+		"globals": "^16.0.0",
+		"prettier": "^3.4.2",
+		"prettier-plugin-svelte": "^3.3.3",
+		"prettier-plugin-tailwindcss": "^0.6.11",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
+		"tailwind-merge": "^3.0.2",
+		"tailwindcss": "^4.0.0",
+		"typescript": "^5.0.0",
+		"typescript-eslint": "^8.20.0",
+		"vite": "^6.0.0"
+	},
+	"pnpm": {
+		"onlyBuiltDependencies": [
+			"esbuild"
+		]
+	},
+	"dependencies": {
+		"@sanity/client": "^6.28.2",
+		"@sanity/image-url": "^1.1.0"
 	}
 }
