@@ -1,41 +1,88 @@
-import {generateDocumentStructure, generateSingletonStructure} from '../utils/desk'
+import {generateDeskStructure} from '../utils/desk'
 import type {StructureBuilder} from 'sanity/structure'
 
-const documents = [] as any[]
+interface DocumentItem {
+  title: string
+  type: string
+  icon?: () => string
+  singleton?: boolean
+  divider?: boolean
+}
+
+interface GroupDocument {
+  title: string
+  type: 'group'
+  icon?: () => string
+  items: DocumentItem[]
+  divider?: boolean
+}
+
+type Document = DocumentItem | GroupDocument
+
+const documents: Document[] = [
+  {
+    title: 'Pages',
+    type: 'group',
+    icon: () => '📚',
+    divider: true,
+    items: [
+      {
+        title: 'Home',
+        type: 'homePage',
+        icon: () => '🏠',
+        singleton: true,
+      },
+      {
+        title: 'Info',
+        type: 'info',
+        icon: () => 'ℹ️',
+        singleton: true,
+        divider: true,
+      },
+      {
+        title: 'Page',
+        type: 'page',
+        icon: () => '📄',
+      },
+    ],
+  },
+  {
+    title: 'Projects',
+    type: 'project',
+    icon: () => '📁',
+    divider: true,
+  },
+  {
+    title: 'Site',
+    type: 'site',
+    icon: () => '🌐',
+    singleton: true,
+  },
+  {
+    title: 'Settings',
+    type: 'settings',
+    icon: () => '⚙️',
+    singleton: true,
+  },
+]
 
 const DOCUMENT_TYPES_IN_STRUCTURE = [
-  `media.tag`,
-  `homePage`,
-  `project`,
-  `settings`,
-  ...documents.map((document) => document.type),
+  'media.tag',
+  ...documents.flatMap((document) => {
+    if (document.type === 'group') {
+      return (document as GroupDocument).items.map((item) => item.type)
+    }
+
+    return document.type
+  }),
 ] as string[]
 
 export const deskStructure = (S: StructureBuilder) =>
   S.list()
-    .title(`Content`)
+    .title('Content')
     .items([
-      ...documents.map((document) => generateDocumentStructure(S, document)),
-      S.divider(),
-      generateSingletonStructure(S, {
-        title: `Home`,
-        type: `homePage`,
-        icon: () => `🏠`,
-      }),
-      S.divider(),
-      generateDocumentStructure(S, {
-        title: `Projects`,
-        type: `project`,
-        icon: () => `📁`,
-      }),
-      S.divider(),
-      generateSingletonStructure(S, {
-        title: `Settings`,
-        type: `settings`,
-        icon: () => `⚙️`,
-      }),
-      ...S.documentTypeListItems().filter((listItem) => {
-        const id = listItem.getId()
-        return id && !DOCUMENT_TYPES_IN_STRUCTURE.includes(id)
-      }),
+      ...generateDeskStructure(S, documents),
+      ...S.documentTypeListItems().filter(
+        (listItem) => !DOCUMENT_TYPES_IN_STRUCTURE.includes(listItem.getId() || ''),
+      ),
     ])
