@@ -3,10 +3,10 @@ import {defineField, defineType} from 'sanity'
 import sections from '../../utils/sections'
 
 export default defineType({
-  name: 'homePage',
-  title: 'Home',
+  name: 'info',
+  title: 'Info Page',
   type: 'document',
-  icon: () => `🏠`,
+  icon: () => `ℹ️`,
   groups: [
     {
       name: 'content',
@@ -25,11 +25,21 @@ export default defineType({
       validation: (Rule) => Rule.required(),
       group: 'content',
     }),
+    defineField({
+      name: 'slug',
+      title: 'Slug',
+      type: 'slug',
+      options: {
+        source: 'title',
+      },
+      validation: (Rule) => Rule.required(),
+      group: 'content',
+    }),
     sections,
     defineField({
       name: 'seo',
       title: 'SEO',
-      type: 'seo.home',
+      type: 'seo.info',
       group: 'seo',
     }),
   ],
