@@ -1,7 +1,7 @@
 import {defineField, defineType} from 'sanity'
 
 export default defineType({
-  name: 'seo.settings',
+  name: 'seo.site',
   title: 'SEO',
   type: 'object',
   fields: [
@@ -19,11 +19,11 @@ export default defineType({
         Rule.max(150).warning('Longer descriptions may be truncated by search engines'),
     }),
     defineField({
-      name: `keywords`,
-      title: `Keywords`,
-      type: `array`,
-      of: [{type: `string`}],
-      options: {layout: `tags`},
+      name: 'keywords',
+      title: 'Keywords',
+      type: 'array',
+      of: [{type: 'string'}],
+      options: {layout: 'tags'},
     }),
     defineField({
       name: 'favicon',
@@ -32,10 +32,11 @@ export default defineType({
       description: 'Image should have a 1:1 aspect ratio, no larger that 512x512 pixels',
     }),
     defineField({
-      name: `image`,
-      title: `Image`,
-      type: `image`,
-      description: `Used for both search engine results and social cards. Image should have a 16:9 aspect ratio. eg. 1200 x 675 pixels`,
+      name: 'image',
+      title: 'Image',
+      type: 'image',
+      description:
+        'Used for both search engine results and social cards. Image should have a 16:9 aspect ratio. eg. 1200 x 675 pixels',
     }),
   ],
   options: {
